Fleet + Add ship and shipStatus

+++ app/src/main/java/florian/com/outerspacemanager/outerspacemanager/OnGridViewSpaceShipChildrenMaxClick.java
@@ -3,5 +3,5 @@ package florian.com.outerspacemanager.outerspacemanager;
 import android.view.View;
 
 public interface OnGridViewSpaceShipChildrenMaxClick {
-    public void OnClick(Double amount, View v);
+    public void OnClick(int amount, View v);
 }
